Created different components to separate logic and clean up StripeForm
