+++ client/src/components/StripeCheckoutForm/AmountCredits.js
@@ -1,0 +1,35 @@
+import React from 'react';
+
+import { withStyles } from '@material-ui/core/styles';
+import TextField from '@material-ui/core/TextField';
+
+import styles from './styles';
+
+const AmountCredits = props => {
+  const { classes, onChangeHandler, value, disabled } = props;
+  return (
+    <TextField
+      InputProps={{
+        inputProps: {
+          className: classes.textField
+        }
+      }}
+      InputLabelProps={{
+        className: classes.label
+      }}
+      id="standard-number"
+      label="How many credits do you want?"
+      name="quantity"
+      disabled={disabled}
+      value={value}
+      onChange={onChangeHandler}
+      className={classes.textField}
+      InputLabelProps={{
+        className: classes.label
+      }}
+      margin="normal"
+    />
+  );
+};
+
+export default withStyles(styles)(AmountCredits);
